Cleanup of traversers, context
This still needs a lot of work. After trying various things, I've given
up on this for now. The end goal is to eliminate one of the traversal
implementations. The stack based one is necessary for backtracking to
work, but adding an option to get the path index slows everything down.
The recursive algorithm is faster in all cases, but cannot be copied and
stored for backtracking. For now, I've removed all but one use of the
stack based version.

Also cleaned up the context interface, and added `get` and `index`
methods. The name and interface of these will probably change in the
near future.